Move PostCardData to its own file
